requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
anthropic==0.7.0
requests==2.31.0
BeautifulSoup4==4.12.2
stripe==7.3.0
PDFKit==1.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
email-validator==2.0.0
celery==5.3.4
redis==5.0.1
flask-migrate==4.0.5
